backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
google-generativeai==0.7.2
uvicorn==0.30.3
drf-spectacular==0.27.0
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
